edunexa/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
Werkzeug==3.0.1
google-generativeai==0.3.2
PyPDF2==3.0.1
moviepy==1.0.3
Pillow==10.1.0
gunicorn==21.2.0
python-multipart==0.0.6
bcrypt==4.1.2
jwt==1.3.1
requests==2.31.0
opencv-python==4.8.1.78
pytube==15.0.0
# speech-recognition
pydub==0.25.1
langchain==0.1.0
streamlit==1.29.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.12.2
scikit-learn==1.3.2
# EOF